[#8] refactor: move table-related components to table directory
